new ssd branch from weiliu89/caffe

+++ scripts/travis/install-deps.sh
@@ -12,6 +12,7 @@ apt-get install -y --no-install-recommends \
   libboost-python-dev \
   libboost-system-dev \
   libboost-thread-dev \
+  libboost-regex-dev \
   libgflags-dev \
   libgoogle-glog-dev \
   libhdf5-serial-dev \
@@ -40,25 +41,30 @@ else
     python3-skimage
 
   # build Protobuf3 since it's needed for Python3
-  echo "Building protobuf3 from source ..."
+  PROTOBUF3_DIR=~/protobuf3
   pushd .
-  PROTOBUF3_DIR=~/protobuf3-build
-  rm -rf $PROTOBUF3_DIR
-  mkdir $PROTOBUF3_DIR
+  if [ -d "$PROTOBUF3_DIR" ] && [ -e "$PROTOBUF3_DIR/src/protoc" ]; then
+    echo "Using cached protobuf3 build ..."
+    cd $PROTOBUF3_DIR
+  else
+    echo "Building protobuf3 from source ..."
+    rm -rf $PROTOBUF3_DIR
+    mkdir $PROTOBUF3_DIR
 
-  # install some more dependencies required to build protobuf3
-  apt-get install -y --no-install-recommends \
-    curl \
-    dh-autoreconf \
-    unzip
+    # install some more dependencies required to build protobuf3
+    apt-get install -y --no-install-recommends \
+      curl \
+      dh-autoreconf \
+      unzip
 
-  wget https://github.com/google/protobuf/archive/3.0.x.tar.gz -O protobuf3.tar.gz
-  tar -xzf protobuf3.tar.gz -C $PROTOBUF3_DIR --strip 1
-  rm protobuf3.tar.gz
-  cd $PROTOBUF3_DIR
-  ./autogen.sh
-  ./configure --prefix=/usr
-  make --jobs=$NUM_THREADS
+    wget https://github.com/google/protobuf/archive/3.0.x.tar.gz -O protobuf3.tar.gz
+    tar -xzf protobuf3.tar.gz -C $PROTOBUF3_DIR --strip 1
+    rm protobuf3.tar.gz
+    cd $PROTOBUF3_DIR
+    ./autogen.sh
+    ./configure --prefix=/usr
+    make --jobs=$NUM_THREADS
+  fi
   make install
   popd
 fi
